Move CheckWooPluginPushNotificationsSupport to notifications/push and consider Woo plugin compatibility in push token registration check
