+++ WooCommerce/src/main/kotlin/com/woocommerce/android/notifications/push/CheckWooPluginPushNotificationsSupport.kt
@@ -1,18 +1,24 @@
-package com.woocommerce.android.ui.pushnotifications
+package com.woocommerce.android.notifications.push
 
 import com.woocommerce.android.extensions.isVersionAtLeast
 import com.woocommerce.android.util.FetchActiveWCPluginVersion
+import com.woocommerce.android.util.GetWooCorePluginCachedVersion
 import javax.inject.Inject
 
 class CheckWooPluginPushNotificationsSupport @Inject constructor(
-    private val fetchActiveWCPluginVersion: FetchActiveWCPluginVersion
+    private val fetchActiveWCPluginVersion: FetchActiveWCPluginVersion,
+    private val getWooCorePluginCachedVersion: GetWooCorePluginCachedVersion
 ) {
     companion object {
         const val PUSH_NOTIFICATIONS_MIN_WC_VERSION = "10.6.0"
     }
 
-    suspend operator fun invoke(): Result {
-        val wcVersion = fetchActiveWCPluginVersion() ?: return Result.Error
+    suspend operator fun invoke(forceRefresh: Boolean = true): Result {
+        val wcVersion = if (forceRefresh) {
+            fetchActiveWCPluginVersion()
+        } else {
+            getWooCorePluginCachedVersion()
+        } ?: return Result.Error
 
         return if (wcVersion.isVersionAtLeast(PUSH_NOTIFICATIONS_MIN_WC_VERSION)) {
             Result.Compatible
